updated ID types for user

+++ OnlineRailwayReservation/src/main/java/com/sunbeam/dto/UserDTO.java
@@ -9,10 +9,10 @@ import com.sunbeam.entity.User;
 import org.springframework.beans.BeanUtils;
 
 public class UserDTO {
-	private int userId;
+	private Long userId;
 	private String email;
 	private String password;
-	private int age;
+	private Integer age;
 	private String gender;
 	private String mobile;
 	private String state;
@@ -23,7 +23,7 @@ public class UserDTO {
 
 	}
 
-	public UserDTO(int userId, String email, String password, int age, String gender, String mobile, String state,
+	public UserDTO(Long userId, String email, String password, int age, String gender, String mobile, String state,
 			String city, String role) {
 		super();
 		this.userId = userId;
@@ -37,11 +37,11 @@ public class UserDTO {
 		this.role = role;
 	}
 
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
@@ -61,11 +61,11 @@ public class UserDTO {
 		this.password = password;
 	}
 
-	public int getAge() {
+	public Integer getAge() {
 		return age;
 	}
 
-	public void setAge(int age) {
+	public void setAge(Integer age) {
 		this.age = age;
 	}
